Separate rollup watch and bundle modules into separate files
